Switch web framework from Cuba to Roda
Roda is a fork of Cuba.  Here are the advantages I see to this change:

1) Uses terminal routes.  One of the reasons behind the fork is that
by default, routes in Cuba are not terminal.  So the following
requests are currently handled the same by Tabula:

POST /upload
POST /upload/with/additional/stuff

Basically, you can stick whatever you want at the end of a path, and
Cuba will just ignore it.  Roda has built in support for terminal
routes, so adding stuff at the end of the path will result in a
404, like practically all websites.

2) Comes with a JSON plugin.  This simplifies JSON handling, and makes
it so you don't have to set the Content-Type to application/json and
call .to_json manually.  You just have the route block return an array
or hash, and Roda will automatically convert it to json for you.

While I was converting the app from Cuba to Roda, I noticed there were
some parts that could derive more benefit from the routing tree.  One
example is TabulaDebug, where I moved a some duplicate code in every
route up to the enclosing branch, so the behavior is shared by all of
the routes under the branch.

Some things get a little more verbose with the switch, such as the
second argument to Roda#view is a general options hash instead of
specific to local variables.  In general that could be made simpler by
passing data to the views implicitly using instance variables instead
of local variables.

Some other changes:

1) require 'tilt/erb' explicitly.  This avoids some warnings printed
by tilt, which should be the case for both Roda and Cuba.

2) require tabula_debug and tabula_job_progress outside of the
routing tree, so they aren't required on every request that uses them.